\name{oauth}
\alias{oauth}
\title{oauth package}
\description{This is a simple implementation of an Oauth 1.0 client for the R statistical programming language. External dependencies are limited to RCurl and digest. Syntax is (largely) simple.}
\usage{Authorize(); MakeRequest()}
\arguments{}
\seealso{}
\examples{
  \dontrun{demo("term-ie_demo")}
}
